Move visitor utils to `rustc_ast_ir`
